+++ src/list.ts
@@ -8,7 +8,7 @@ import { createItemTemplate } from "./utils/createItemTemplate.js";
 import { itemIteratorNext } from "./utils/listItemIterator.js";
 import { generateID } from "./utils/generateID.js";
 import { getURLParam } from "./utils/getURLParam.js";
-import { getListItems } from "./utils/getListItems.js";
+import { addListItem, getListItems } from "./utils/localStoragAPI.js";
 
 const MAX_SUGGESTED_ITEMS = 4;
 const IS_INDEX_PAGE = false;
@@ -48,6 +48,7 @@ const list = InitializeList({
   listID,
   onAddItem: (item) => {
     console.log("item added...", item);
+    listID && addListItem(listID, item);
   },
   ondeleteItem: (itemID) => {
     console.log("item deleted...", itemID);
@@ -90,6 +91,7 @@ localList?.forEach((item) => {
     list: list,
     actionButtonType,
     showInputDefault: false,
+    // isFromBackEnd: true,
   });
 });
