+++ src/utils/localStoragAPI.ts
@@ -1,0 +1,37 @@
+import { ListItem } from "../types/types";
+
+/**
+ * gets list items with a listID from the local storage
+ * @param listID listID to search for items in the local storage for
+ * @returns items stored in the local storage
+ */
+export function getListItems(listID: string | undefined) {
+  // if listID is undefined, stop here
+  if (!listID) return;
+
+  // if no list exists in local storage stop here
+  const temp = localStorage.getItem(`list--${listID}`);
+  if (!temp) return;
+
+  // return list parsed to listItem array
+  return JSON.parse(temp) as ListItem[];
+}
+
+/**
+ * adds an item to a list stored in local storage
+ * @param listID listID to add item to
+ * @param item item to add to list
+ * @returns void
+ */
+export function addListItem<T>(listID: string, item: T) {
+  // if no list exists in local storage, stop here
+  const temp = localStorage.getItem(`list--${listID}`);
+  if (!temp) return;
+
+  // add item to list
+  const list = JSON.parse(temp) as T[];
+  list.push(item);
+
+  // update local storage
+  localStorage.setItem(`list--${listID}`, JSON.stringify(list));
+}
